+++ sources/Deltares.Probabilistic/Combine/HohenbichlerNumIntCombiner.cpp
@@ -23,7 +23,7 @@
 
 #include "CombineElements.h"
 #include "HohenbichlerNumInt.h"
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 namespace Deltares::Reliability
 {
