+++ sources/Deltares.Probabilistic/Logging/Message.cpp
@@ -20,7 +20,7 @@
 // All rights reserved.
 //
 #include "Message.h"
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 namespace Deltares::Logging
 {
