+++ sources/Deltares.Probabilistic/Math/CholeskyDecomp.cpp
@@ -21,7 +21,7 @@
 //
 #include <cmath>
 #include "Matrix.h"
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 namespace Deltares::Numeric
 {
