+++ sources/Deltares.Probabilistic/Math/Matrix.h
@@ -23,7 +23,7 @@
 #include <iosfwd>
 #include <vector>
 #include "Vector1D.h"
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 namespace Deltares::Numeric
 {
