+++ sources/Deltares.Probabilistic/Math/NumericSupport.cpp
@@ -28,7 +28,7 @@
 #include <sstream>
 
 #include "Constants.h"
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 namespace Deltares::Numeric
 {
