+++ sources/Deltares.Probabilistic/Math/SpecialFunctions.cpp
@@ -23,7 +23,7 @@
 
 #include <cmath>
 
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 namespace Deltares::Numeric
 {
