+++ sources/Deltares.Probabilistic/Math/Vector1D.h
@@ -23,7 +23,7 @@
 #include <iosfwd>
 #include <vector>
 #include <initializer_list>
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 namespace Deltares::Numeric
 {
