+++ sources/Deltares.Probabilistic/Model/GradientCalculator.cpp
@@ -26,7 +26,7 @@
 #include "GradientSettings.h"
 #include "Sample.h"
 #include "ModelRunner.h"
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 using namespace Deltares::Reliability;
