+++ sources/Deltares.Probabilistic/Model/GradientSettings.cpp
@@ -21,7 +21,7 @@
 //
 
 #include "GradientSettings.h"
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 namespace Deltares::Models
 {
