+++ sources/Deltares.Probabilistic/Server/ProjectEntries.cpp
@@ -20,7 +20,7 @@
 // All rights reserved.
 //
 #include "ProjectEntries.h"
-#include "../Utils/probLibException.h"
+#include "../Utils/ProbabilisticLibraryException.h"
 
 namespace Deltares::Server
 {
